feat: finish modules tasks
